requirements.txt
Django==3.0.5
asgiref==3.2.7
django-enum-choices==2.1.1
djangorestframework==3.11.0
pip==19.2.3
pytz==2019.3
setuptools==40.8.0
sqlparse==0.3.1
virtualenv==16.7.5
wheel==0.33.6
faker==4.0.2
django-allauth==0.41.0
django-rest-auth==0.9.5
mysqlclient==1.4.6
django-heroku
